fix(utils): move name normalization function to its own utils file
