+++ api/src/app/utils/name_utils.py
@@ -1,0 +1,3 @@
+def normalize_resource_name(name: str) -> str:
+    """Remove problematic characters from user-supplied resource names."""
+    return name.strip().replace(" ", "")
